Level4: Assets, Fog Of War, XP Orbs, Tiles added; tweaked Lighting
Added a bunch of assets, honestly I don't anymore is a good idea despite some empty gaps between tree groups and edges of pathways since tris are nearing 1 mil in a frame. I added the Fog of War with the assets in the triggers and the tiles + orbs in the scene. Just need to do playtests and adjustments with said playtests